Copy build script outputs to dist directory via xtask
This is so much more complicated than it needs to be. The complexity
comes from the build.rs OUT_DIR being randomly generated, and there's
not an easy way to move files out of there or even reference where the
output directory is. Unfortunately, putting things like the completions
files and man pages in the OUT_DIR is the "proper" way to do things, and
other projects have all sorts of ugly workarounds.

There is an unstable nightly flag `--out-dir` that I thought would
solve the problem, but it only copies the build binary and not the build
script outputs. The workaround here is to parse the json cargo messages
directly to grab the OUT_DIR, and then recursively copy any asset files
ourselves.

See:
- https://github.com/rust-lang/cargo/issues/6790
- https://github.com/benesch/cargo-out-dir/blob/master/src/main.rs

+++ xtask/src/dist.rs
@@ -9,8 +9,15 @@ use nanoserde::DeJson;
 use xshell::Shell;
 
 use crate::commands::cargo_cmd;
-use crate::utils::{project_root, verbose_cd};
+use crate::utils::{copy_dir_to, project_root, verbose_cd};
 use crate::Config;
+
+#[derive(Debug, DeJson)]
+struct CheckMessage {
+    out_dir: Option<String>,
+    package_id: String,
+    reason: String,
+}
 
 #[derive(Debug, DeJson)]
 struct Metadata {
@@ -31,11 +38,13 @@ struct Target {
 }
 
 pub fn dist(config: &Config) -> Result<()> {
-    fs::remove_dir_all(dist_dir())?;
-    fs::create_dir_all(dist_dir())?;
+    if dist_dir().exists() {
+        fs::remove_dir_all(dist_dir())?;
+    }
 
     dist_binary(config)?;
-    // TODO: dist_manpage()?;
+    dist_build_script_outputs(config)?;
+    // TODO: dist_docs()?;
 
     Ok(())
 }
@@ -44,17 +53,13 @@ fn dist_binary(config: &Config) -> Result<()> {
     let sh = Shell::new()?;
     verbose_cd(&sh, project_root());
 
-    let cmd_option = cargo_cmd(config, &sh);
-    if let Some(cmd) = cmd_option {
-        let args = vec!["build", "--profile", "production", "--bins"];
-        cmd.args(args).run()?;
-    }
-
     let binaries = project_binaries(config)?;
+
     for binary in &binaries {
-        if binary == "xtask" {
-            eprintln!("Ignoring xtask binary");
-            continue;
+        let cmd_option = cargo_cmd(config, &sh);
+        if let Some(cmd) = cmd_option {
+            let args = vec!["build", "--profile", "production", "--bin", binary];
+            cmd.args(args).run()?;
         }
 
         let binary_filename = if cfg!(target_os = "windows") {
@@ -62,11 +67,58 @@ fn dist_binary(config: &Config) -> Result<()> {
         } else {
             binary.to_string()
         };
-        let src = release_dir().join(&binary_filename);
-        let dest = dist_dir().join(&binary_filename);
+        let src = production_dir().join(&binary_filename);
+
+        // Destination: target/dist/binary/binary_filename
+        let dest_dir = dist_dir().join(binary);
+        fs::create_dir_all(&dest_dir)?;
+        let dest = dest_dir.join(&binary_filename);
 
         eprintln!("Copying {} to {}", src.display(), dest.display());
         fs::copy(&src, &dest)?;
+    }
+
+    Ok(())
+}
+
+fn dist_build_script_outputs(config: &Config) -> Result<()> {
+    let sh = Shell::new()?;
+    verbose_cd(&sh, project_root());
+
+    let binaries = project_binaries(config)?;
+
+    for binary in &binaries {
+        let cmd_option = cargo_cmd(config, &sh);
+        if let Some(cmd) = cmd_option {
+            eprintln!(
+                "$ cargo check --profile production --message-format=json --quiet --bin {binary}"
+            );
+            let args = vec![
+                "check",
+                "--profile",
+                "production",
+                "--message-format=json",
+                "--quiet",
+                "--bin",
+                binary,
+            ];
+            let output = cmd.args(args).output()?;
+
+            let check_json = String::from_utf8(output.stdout)?;
+            let out_dir_option = get_out_dir(&check_json, binary)?;
+
+            if let Some(out_dir) = out_dir_option {
+                let absolute_out_dir = PathBuf::from(out_dir);
+                let src_dir = absolute_out_dir.strip_prefix(sh.current_dir())?;
+
+                // Destination: target/dist/binary/
+                let dest_dir = dist_dir().join(binary);
+                fs::create_dir_all(&dest_dir)?;
+
+                eprintln!("Copying {}/* to {}/", src_dir.display(), dest_dir.display());
+                copy_dir_to(src_dir, &dest_dir)?;
+            };
+        }
     }
 
     Ok(())
@@ -86,8 +138,8 @@ fn project_binaries(config: &Config) -> Result<Vec<String>> {
 
         for package in metadata.packages {
             for target in &package.targets {
-                if target.kind.contains(&"bin".to_string()) {
-                    eprintln!("{package:?}");
+                if target.name != "xtask" && target.kind.contains(&"bin".to_string()) {
+                    // eprintln!("{package:?}");
                     binaries.push(target.name.clone());
                 }
             }
@@ -97,12 +149,25 @@ fn project_binaries(config: &Config) -> Result<Vec<String>> {
     Ok(binaries)
 }
 
+fn get_out_dir(check_json: &str, binary: &str) -> Result<Option<String>> {
+    for line in check_json.lines() {
+        let msg: CheckMessage = DeJson::deserialize_json(line)?;
+        if msg.reason == "build-script-executed"
+            && msg.package_id.starts_with(binary)
+            && msg.out_dir.is_some()
+        {
+            return Ok(msg.out_dir);
+        }
+    }
+    Ok(None)
+}
+
 fn dist_dir() -> PathBuf {
     target_dir().join("dist")
 }
 
-fn release_dir() -> PathBuf {
-    target_dir().join("release")
+fn production_dir() -> PathBuf {
+    target_dir().join("production")
 }
 
 fn target_dir() -> PathBuf {
